requirements.txt
-e .
aws_cdk.aws_autoscaling
aws_cdk.aws_elasticloadbalancingv2
aws_cdk.aws_elasticloadbalancingv2_targets
aws_cdk.aws_ec2
aws_cdk.aws_iam
